+++ NEWS.md
@@ -4,6 +4,8 @@
 - In case both an error and a warning are raised by the function being
   fuzzed, we now report the error; previously the warning was reported
   because of how `tryCatch()` is implemented.
+- The `length()` S3 method for objects of class `cbtf` now counts the total
+  number of tests performed, instead of the number of inputs tested.
 
 # CBTF 0.4.0 (2025-07-31)